pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "codd-dev"
version = "2.14.0"
description = "CoDD: Coherence-Driven Development — cross-artifact change impact analysis"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Yohei Watanabe", email = "yohey-w@users.noreply.github.com" },
]
keywords = ["change-impact", "dependency-graph", "software-engineering", "claude-code", "plugin"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "pyyaml>=6.0",
    "click>=8.0",
    "jinja2>=3.1.0",
    "tomli>=2.0.1; python_version < '3.11'",
    "watchdog>=4.0.0",
    "websocket-client>=1.8.0",
]

[project.optional-dependencies]
ai = []
api-parsers = [
    "graphql-core>=3.2.0",
]
infra = [
    "python-hcl2>=7.0.0",
]
tree-sitter = [
    "tree-sitter>=0.25.0,<0.26.0",
    "tree-sitter-python>=0.25.0,<0.26.0",
    "tree-sitter-typescript>=0.23.0,<0.24.0",
]
scan = [
    "tree-sitter>=0.22",
    "tree-sitter-python>=0.22",
    "tree-sitter-typescript>=0.22",
    "tree-sitter-java>=0.22",
]
mcp = []

[project.scripts]
codd = "codd.cli:main"

[project.urls]
Homepage = "https://github.com/yohey-w/codd-dev"
Repository = "https://github.com/yohey-w/codd-dev"
Issues = "https://github.com/yohey-w/codd-dev/issues"

[tool.hatch.build]
include = [
    "codd/**/*.py",
    "codd/**/*.md",
    "codd/**/*.yaml",
    "codd/hooks/recipes/*.json",
    "codd/hooks/recipes/*.sh",
    "codd/templates/**/*.j2",
    "codd/templates/**/*.tmpl",
    "codd/hooks/pre-commit",
    "codd_plugins/**/*.yaml",
    "codd_plugins/**/*.md",
    "README.md",
    "LICENSE",
]

[tool.hatch.build.targets.wheel]
packages = ["codd"]
